+++ frontend/appflowy_flutter/lib/user/presentation/screens/workspace_start_screen/mobile_workspace_start_screen.dart
@@ -7,6 +7,7 @@ import 'package:flowy_infra_ui/flowy_infra_ui.dart';
 import 'package:flowy_infra_ui/widget/error_page.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';
+import 'package:go_router/go_router.dart';
 
 // TODO(yijing): needs refactor when multiple workspaces are supported
 class MobileWorkspaceStartScreen extends StatefulWidget {
@@ -139,6 +140,5 @@ class _MobileWorkspaceStartScreenState
 // same method as in desktop
 void _popToWorkspace(BuildContext context, WorkspacePB workspace) {
   context.read<WorkspaceBloc>().add(WorkspaceEvent.openWorkspace(workspace));
-
-  Navigator.of(context).pop(workspace.id);
+  context.pop(workspace.id);
 }